break out logic to lib.js
